style: update style up searchable texts
